Flex_report/Flex_report/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.10.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
boto3==1.26.101
botocore==1.29.101
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
config==0.5.1
Cython==0.29.35
Deprecated==1.2.13
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Flask==2.2.3
Flask-APScheduler==1.12.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
logger-setup==0.17
MarkupSafe==2.1.2
moncli==2.0.14
monday==1.3.3
multidict==6.0.4
numpy==1.24.2
openpyxl==3.1.2
outcome==1.2.0
pandas==1.5.3
pdfkit==1.0.0
pyarrow==12.0.1
pycodestyle==2.10.0
pycountry==22.3.5
PyJWT==2.6.0
pylance==0.4.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
s3transfer==0.6.0
schematics==2.1.1
selenium==4.10.0
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.8
starkbank-ecdsa==2.2.0
stripe==5.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
twilio==8.1.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
undetected-chromedriver==3.5.0
urllib3==1.26.15
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.1
